requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
boto3==1.34.162
celery==5.4.0
redis==5.0.7
httpx==0.27.2
jinja2==3.1.4
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
loguru==0.7.2
markdown2==2.5.0
PyYAML==6.0.2
pytest==8.3.2
black==24.8.0
